Add ability to update key

+++ build_usbkey.sh
@@ -10,12 +10,19 @@ target_volume_name="NEO4J"
 blank_volume="$volumes_dir/$blank_volume_name"
 target_volume="$volumes_dir/$target_volume_name"
 
+if [ -d "$blank_volume" ]; then
+  set -x
+  diskutil renameVolume "$blank_volume" "$target_volume_name"
+  sleep 5
+  set +x
+fi
+if ! [ -d "$target_volume" ]; then
+  echo "Can't find \"$target_volume\"" >&2
+  exit 1
+fi
 set -x
-
-diskutil renameVolume "$blank_volume" "$target_volume_name"
-sleep 5
 mdutil -i off "$target_volume"
 sleep 3
-rsync -avP --delete "$dist_dir/USBKEY/" "$target_volume/"
+rsync -avP --size-only --exclude '.*' --delete --delete-excluded "$dist_dir/USBKEY/" "$target_volume/"
 find "$target_volume" -name '.*' -delete
 diskutil eject "$target_volume"
